// frontend/src/lib/api/generated/model/supplierRead.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Construction ERP API
 * OpenAPI spec version: 0.1.0
 */

export interface SupplierRead {
  /**
     * @minLength 1
     * @maxLength 200
     */
  name: string;
  contact_name?: string | null;
  email?: string | null;
  phone?: string | null;
  address?: string | null;
  tax_id?: string | null;
  categories?: string | null;
  notes?: string | null;
  id: string;
  is_active: boolean;
  created_at: string;
}